refactor(duty-tracker): move executors to a separate module
