+++ crates/duty-tracker/src/contract_manager.rs
@@ -2,7 +2,7 @@
 //! and responding to chain events and operator p2p network messages according to the Strata Bridge
 //! protocol rules.
 use std::{
-    collections::{BTreeMap, BTreeSet, HashSet},
+    collections::{BTreeMap, BTreeSet},
     fmt::Debug,
     sync::Arc,
     time::Duration,
@@ -10,44 +10,20 @@ use std::{
 };
 
 use alpen_bridge_params::prelude::{ConnectorParams, PegOutGraphParams, StakeChainParams};
-use bdk_wallet::{miniscript::ToPublicKey, Wallet};
-use bitcoin::{
-    hashes::{sha256, Hash},
-    hex::DisplayHex,
-    sighash::{Prevouts, SighashCache},
-    taproot, Block, FeeRate, Network, OutPoint, Psbt, TapSighashType, Transaction, Txid,
-};
+use bdk_wallet::miniscript::ToPublicKey;
+use bitcoin::{hashes::Hash, hex::DisplayHex, Block, Network, OutPoint, Transaction, Txid};
 use bitcoind_async_client::{client::Client as BitcoinClient, traits::Reader};
-use bitvm::chunk::api::{NUM_HASH, NUM_PUBS, NUM_U256};
 use btc_notify::client::BtcZmqClient;
-use futures::{
-    future::{join3, join_all},
-    StreamExt,
-};
-use musig2::{
-    secp256k1::{self, schnorr, Message},
-    PartialSignature, PubNonce,
-};
-use operator_wallet::{FundingUtxo, OperatorWallet};
+use futures::StreamExt;
+use operator_wallet::OperatorWallet;
 use secret_service_client::SecretServiceClient;
 use secret_service_proto::v1::traits::*;
 use strata_bridge_db::persistent::sqlite::SqliteDb;
 use strata_bridge_p2p_service::MessageHandler;
-use strata_bridge_primitives::{
-    build_context::BuildContext, operator_table::OperatorTable, scripts::taproot::TaprootWitness,
-};
-use strata_bridge_stake_chain::{
-    prelude::StakeTx, stake_chain::StakeChainInputs, transactions::stake::StakeTxData,
-};
-use strata_bridge_tx_graph::{
-    pog_musig_functor::PogMusigF,
-    transactions::{deposit::DepositTx, prelude::CovenantTx},
-};
+use strata_bridge_primitives::operator_table::OperatorTable;
+use strata_bridge_tx_graph::transactions::{deposit::DepositTx, prelude::CovenantTx};
 use strata_p2p::{self, commands::Command, events::Event, swarm::handle::P2PHandle};
-use strata_p2p_types::{
-    P2POperatorPubKey, Scope, SessionId, StakeChainId, Wots128PublicKey, Wots256PublicKey,
-    WotsPublicKeys,
-};
+use strata_p2p_types::{P2POperatorPubKey, Scope, SessionId, StakeChainId};
 use strata_p2p_wire::p2p::v1::{GetMessageRequest, GossipsubMsg, UnsignedGossipsubMsg};
 use strata_primitives::params::RollupParams;
 use strata_state::{bridge_state::DepositState, chain_state::Chainstate};
@@ -56,18 +32,17 @@ use tokio::{
     task::{self, JoinHandle},
     time,
 };
-use tracing::{error, info, trace, warn};
+use tracing::{debug, error, info, trace, warn};
 
 use crate::{
     contract_persister::ContractPersister,
     contract_state_machine::{
         ContractEvent, ContractSM, ContractState, DepositSetup, FulfillerDuty, OperatorDuty,
-        TransitionErr,
     },
     errors::{ContractManagerErr, StakeChainErr},
-    executors::*,
+    executors::prelude::*,
     predicates::{deposit_request_info, parse_strata_checkpoint},
-    s2_session_manager::{MusigSessionErr, MusigSessionManager},
+    s2_session_manager::MusigSessionManager,
     stake_chain_persister::StakeChainPersister,
     stake_chain_state_machine::StakeChainSM,
     tx_driver::TxDriver,
@@ -374,20 +349,20 @@ impl Drop for ContractManager {
 }
 
 /// The handles required by the duty tracker to execute duties.
-struct OutputHandles {
-    wallet: RwLock<OperatorWallet>,
-    msg_handler: MessageHandler,
-    s2_session_manager: MusigSessionManager,
-    tx_driver: TxDriver,
-    db: SqliteDb,
+pub(super) struct OutputHandles {
+    pub(super) wallet: RwLock<OperatorWallet>,
+    pub(super) msg_handler: MessageHandler,
+    pub(super) s2_session_manager: MusigSessionManager,
+    pub(super) tx_driver: TxDriver,
+    pub(super) db: SqliteDb,
 }
 
 /// The actual state that is being tracked by the [`ContractManager`].
 #[derive(Debug)]
-struct ExecutionState {
-    active_contracts: BTreeMap<Txid, ContractSM>,
-    claim_txids: BTreeMap<Txid, Txid>,
-    stake_chains: StakeChainSM,
+pub(super) struct ExecutionState {
+    pub(super) active_contracts: BTreeMap<Txid, ContractSM>,
+    pub(super) claim_txids: BTreeMap<Txid, Txid>,
+    pub(super) stake_chains: StakeChainSM,
 }
 
 /// pub(super) The proxy for the state being tracked by the [`ContractManager`].
@@ -1226,11 +1201,10 @@ async fn execute_duty(
     duty: OperatorDuty,
 ) -> Result<(), ContractManagerErr> {
     let OutputHandles {
-        wallet,
         msg_handler,
         s2_session_manager,
-        tx_driver,
         db,
+        ..
     } = &*output_handles;
 
     match duty {
@@ -1251,17 +1225,92 @@ async fn execute_duty(
         } => {
             handle_publish_deposit_setup(
                 &cfg,
-                &s2_session_manager.s2_client,
-                db,
-                wallet,
-                msg_handler,
-                tx_driver,
+                output_handles.clone(),
                 deposit_txid,
                 deposit_idx,
                 stake_chain_inputs,
             )
             .await
         }
+
+        OperatorDuty::PublishRootNonce {
+            deposit_request_txid,
+            witness,
+        } => {
+            handle_publish_root_nonce(
+                &output_handles.s2_session_manager,
+                &output_handles.msg_handler,
+                OutPoint::new(deposit_request_txid, 0),
+                witness,
+            )
+            .await
+        }
+
+        OperatorDuty::PublishGraphNonces {
+            claim_txid,
+            pog_prevouts: pog_inputs,
+            pog_witnesses,
+        } => {
+            handle_publish_graph_nonces(
+                &output_handles.s2_session_manager,
+                &output_handles.msg_handler,
+                claim_txid,
+                pog_inputs,
+                pog_witnesses,
+            )
+            .await
+        }
+
+        OperatorDuty::PublishGraphSignatures {
+            claim_txid,
+            pubnonces,
+            pog_prevouts: pog_outpoints,
+            pog_sighashes,
+        } => {
+            handle_publish_graph_sigs(
+                &output_handles.s2_session_manager,
+                &output_handles.msg_handler,
+                claim_txid,
+                pubnonces,
+                pog_outpoints,
+                pog_sighashes,
+            )
+            .await
+        }
+
+        OperatorDuty::PublishRootSignature {
+            nonces,
+            deposit_request_txid,
+            sighash,
+        } => {
+            handle_publish_root_signature(
+                &cfg,
+                &output_handles.s2_session_manager,
+                &output_handles.msg_handler,
+                nonces,
+                OutPoint::new(deposit_request_txid, 0),
+                sighash,
+            )
+            .await
+        }
+
+        OperatorDuty::PublishDeposit {
+            deposit_tx,
+
+            partial_sigs,
+        } => {
+            handle_publish_deposit(
+                &output_handles.s2_session_manager,
+                &output_handles.tx_driver,
+                deposit_tx,
+                partial_sigs
+                    .into_iter()
+                    .map(|(k, v)| (cfg.operator_table.op_key_to_btc_key(&k).unwrap(), v))
+                    .collect(),
+            )
+            .await
+        }
+
         OperatorDuty::FulfillerDuty(fulfiller_duty) => match fulfiller_duty {
             FulfillerDuty::AdvanceStakeChain {
                 stake_index,
@@ -1322,620 +1371,4 @@ async fn execute_duty(
             Ok(())
         }
     }
-}
-
-async fn handle_publish_stake_chain_exchange(
-    cfg: &ExecutionConfig,
-    s2_client: &SecretServiceClient,
-    db: &SqliteDb,
-    msg_handler: &MessageHandler,
-) -> Result<(), ContractManagerErr> {
-    let pov_idx = cfg.operator_table.pov_idx();
-    let general_key = s2_client
-        .general_wallet_signer()
-        .pubkey()
-        .await?
-        .to_x_only_pubkey();
-
-    if let Some(pre_stake) = db
-        .get_pre_stake(pov_idx)
-        .await
-        .expect("should be able to consult the database")
-    {
-        let stake_chain_id = StakeChainId::from_bytes([0u8; 32]);
-        info!(%stake_chain_id, "broadcasting pre-stake information");
-
-        msg_handler
-            .send_stake_chain_exchange(stake_chain_id, general_key, pre_stake.txid, pre_stake.vout)
-            .await;
-
-        return Ok(());
-    }
-
-    error!("pre-stake information does exist in the database");
-
-    Err(TransitionErr(
-        "pre-stake information missing in the database".to_string(),
-    ))?
-}
-
-async fn finalize_claim_funding_tx(
-    s2_client: &SecretServiceClient,
-    tx_driver: &TxDriver,
-    general_wallet: &Wallet,
-    psbt: Psbt,
-) -> Result<(), ContractManagerErr> {
-    let mut tx = psbt.unsigned_tx;
-    let txins_as_outs = tx
-        .input
-        .iter()
-        .map(|txin| {
-            general_wallet
-                .get_utxo(txin.previous_output)
-                .expect("always have this output because the wallet selected it in the first place")
-                .txout
-        })
-        .collect::<Vec<_>>();
-    let mut sighasher = SighashCache::new(&mut tx);
-    let sighash_type = TapSighashType::All;
-    let prevouts = Prevouts::All(&txins_as_outs);
-    for input_index in 0..txins_as_outs.len() {
-        let sighash = sighasher
-            .taproot_key_spend_signature_hash(input_index, &prevouts, sighash_type)
-            .expect("failed to construct sighash");
-        let signature = s2_client
-            .general_wallet_signer()
-            .sign(&sighash.to_byte_array(), None)
-            .await?;
-
-        let signature = bitcoin::taproot::Signature {
-            signature,
-            sighash_type,
-        };
-        sighasher
-            .witness_mut(input_index)
-            .expect("an input here")
-            .push(signature.to_vec());
-    }
-
-    info!(
-        txid = %tx.compute_txid(),
-        "submitting claim funding tx to the tx driver"
-    );
-    tx_driver
-        .drive(tx, 0)
-        .await
-        .map_err(|e| ContractManagerErr::FatalErr(Box::new(e)))?;
-
-    Ok(())
-}
-
-async fn handle_advance_stake_chain(
-    cfg: &ExecutionConfig,
-    output_handles: Arc<OutputHandles>,
-    stake_index: u32,
-    stake_tx: StakeTx,
-) -> Result<(), ContractManagerErr> {
-    let operator_id = cfg.operator_table.pov_idx();
-    let op_p2p_key = cfg.operator_table.pov_op_key();
-
-    let pre_stake_outpoint = output_handles
-        .db
-        .get_pre_stake(operator_id)
-        .await?
-        .ok_or(StakeChainErr::StakeSetupDataNotFound(op_p2p_key.clone()))?;
-
-    let messages = stake_tx.sighashes();
-    let funds_signature = output_handles
-        .s2_session_manager
-        .s2_client
-        .general_wallet_signer()
-        .sign(messages[0].as_ref(), None)
-        .await?;
-
-    let signed_stake_tx = if stake_index == 0 {
-        // the first stake transaction spends the pre-stake which is locked by the key in the
-        // stake-chain wallet
-        let stake_signature = output_handles
-            .s2_session_manager
-            .s2_client
-            .general_wallet_signer()
-            .sign(messages[1].as_ref(), None)
-            .await?;
-
-        stake_tx.finalize_initial(funds_signature, stake_signature)
-    } else {
-        let pre_image_client = output_handles
-            .s2_session_manager
-            .s2_client
-            .stake_chain_preimages();
-        let OutPoint {
-            txid: pre_stake_txid,
-            vout: pre_stake_vout,
-        } = pre_stake_outpoint;
-        let prev_preimage = pre_image_client
-            .get_preimg(pre_stake_txid, pre_stake_vout, stake_index - 1)
-            .await?;
-        let n_of_n_agg_pubkey = cfg
-            .operator_table
-            .tx_build_context(cfg.network)
-            .aggregated_pubkey();
-        let operator_pubkey = output_handles
-            .s2_session_manager
-            .s2_client
-            .general_wallet_signer()
-            .pubkey()
-            .await?
-            .to_x_only_pubkey();
-        let stake_hash = pre_image_client
-            .get_preimg(pre_stake_txid, pre_stake_vout, stake_index)
-            .await?;
-        let stake_hash = sha256::Hash::hash(&stake_hash);
-        let StakeChainParams { delta, .. } = cfg.stake_chain_params;
-        let prev_connector_s = ConnectorStake::new(
-            n_of_n_agg_pubkey,
-            operator_pubkey,
-            stake_hash,
-            delta,
-            cfg.network,
-        );
-
-        // all the stake transactions except the first one are locked with the general wallet
-        // signer.
-        // this is a caveat of the fact that we only share one x-only pubkey during deposit
-        // setup which is used for reimbursements/cpfp.
-        // so instead of sharing ones, we can just reuse this key (which is part of a taproot
-        // address).
-        let stake_signature = output_handles
-            .s2_session_manager
-            .s2_client
-            .stakechain_wallet_signer()
-            .sign_no_tweak(messages[1].as_ref())
-            .await?;
-
-        stake_tx.finalize(
-            &prev_preimage,
-            funds_signature,
-            stake_signature,
-            prev_connector_s,
-        )
-    };
-
-    let confirm_by = 1;
-    // FIXME: (@Rajil1213) change this to the current block height
-    // once the tx driver's deadline handling is implemented
-    output_handles
-        .tx_driver
-        .drive(signed_stake_tx, confirm_by)
-        .await?;
-
-    Ok(())
-}
-
-async fn handle_publish_graph_nonces(
-    musig: &MusigSessionManager,
-    message_handler: &MessageHandler,
-    claim_txid: Txid,
-    pog_outpoints: PogMusigF<OutPoint>,
-    pog_witnesses: PogMusigF<TaprootWitness>,
-) -> Result<(), ContractManagerErr> {
-    info!(%claim_txid, "executing duty to publish graph nonces");
-
-    let nonces: PogMusigF<PubNonce> = match PogMusigF::transpose_result(
-        pog_outpoints
-            .clone()
-            .zip(pog_witnesses)
-            .map(|(outpoint, witness)| musig.get_nonce(outpoint, witness))
-            .join_all()
-            .await,
-    ) {
-        Ok(res) => res,
-        Err(err) => {
-            match err {
-                MusigSessionErr::SecretServiceClientErr(client_error) => {
-                    warn!(%client_error, "error getting nonces for graph from s2")
-                }
-                MusigSessionErr::SecretServiceNewSessionErr(musig2_new_session_error) => {
-                    // TODO: (@Rajil1213) handle this properly when we known what causes this
-                    error!(
-                        ?musig2_new_session_error,
-                        "error getting nonces for graph from s2"
-                    )
-                }
-                MusigSessionErr::SecretServiceRoundContributionErr(round_contribution_error) => {
-                    // TODO: (@Rajil1213) handle this properly when we known what causes this
-                    error!(%round_contribution_error, "error getting nonces for graph from s2")
-                }
-                MusigSessionErr::SecretServiceRoundFinalizeErr(round_finalize_error) => {
-                    // TODO: (@Rajil1213) handle this properly when we known what causes this
-                    error!(%round_finalize_error, "error getting nonces for graph from s2")
-                }
-                MusigSessionErr::Premature => {
-                    unreachable!("this should never happen unless the stf is wrong")
-                }
-                MusigSessionErr::NotFound(out_point) => {
-                    // this can happen either because the session has already been finalized
-                    // or if the contract is unknown to us
-                    // both of which are okay but we do log it here.
-                    warn!(%out_point, "session outpoint not found");
-                }
-            }
-
-            return Ok(());
-        }
-    };
-
-    info!(%claim_txid, "publishing graph nonces");
-
-    message_handler
-        .send_musig2_nonces(
-            SessionId::from_bytes(claim_txid.to_byte_array()),
-            nonces.pack(),
-        )
-        .await;
-
-    Ok(())
-}
-
-async fn handle_publish_graph_sigs(
-    musig: &MusigSessionManager,
-    message_handler: &MessageHandler,
-    claim_txid: Txid,
-    pubnonces: BTreeMap<secp256k1::PublicKey, PogMusigF<PubNonce>>,
-    pog_outpoints: PogMusigF<OutPoint>,
-    pog_sighashes: PogMusigF<Message>,
-) -> Result<(), ContractManagerErr> {
-    info!(%claim_txid, "executing duty to publish graph signatures");
-
-    // Add all nonces to the musig session manager context.
-    for (pk, graph_nonces) in pubnonces {
-        info!(%pk, "loading nonces");
-
-        PogMusigF::<()>::transpose_result::<MusigSessionErr>(
-            pog_outpoints
-                .clone()
-                .zip(graph_nonces)
-                .map(|(outpoint, nonce)| musig.put_nonce(outpoint, pk.to_x_only_pubkey(), nonce))
-                .join_all()
-                .await,
-        )?;
-    }
-
-    info!(%claim_txid, "getting all partials");
-
-    let partials = PogMusigF::transpose_result(
-        pog_outpoints
-            .zip(pog_sighashes)
-            .map(|(op, sighash)| musig.get_partial(op, sighash))
-            .join_all()
-            .await,
-    )
-    .inspect_err(|e| {
-        error!(
-            %claim_txid,
-            ?e,
-            "failed to get partials for graph signatures"
-        );
-    })?;
-
-    info!(%claim_txid, "publishing graph signatures");
-    debug!(%claim_txid, ?partials, "received all partials from s2");
-
-    message_handler
-        .send_musig2_signatures(
-            SessionId::from_bytes(claim_txid.to_byte_array()),
-            partials.pack(),
-        )
-        .await;
-
-    Ok(())
-}
-
-async fn handle_publish_root_nonce(
-    s2_client: &MusigSessionManager,
-    msg_handler: &MessageHandler,
-    prevout: OutPoint,
-    witness: TaprootWitness,
-) -> Result<(), ContractManagerErr> {
-    let deposit_request_txid = prevout.txid;
-    info!(%deposit_request_txid, "executing duty to publish root nonce");
-
-    let nonce = s2_client.get_nonce(prevout, witness).await?;
-
-    info!(%deposit_request_txid, "publishing root nonce");
-    msg_handler
-        .send_musig2_nonces(
-            SessionId::from_bytes(deposit_request_txid.to_byte_array()),
-            vec![nonce],
-        )
-        .await;
-
-    Ok(())
-}
-
-async fn handle_publish_root_signature(
-    cfg: &ExecutionConfig,
-    s2_client: &MusigSessionManager,
-    msg_handler: &MessageHandler,
-    nonces: BTreeMap<secp256k1::PublicKey, PubNonce>,
-    prevout: OutPoint,
-    sighash: Message,
-) -> Result<(), ContractManagerErr> {
-    let deposit_request_txid = prevout.txid;
-    info!(%deposit_request_txid, "executing duty to publish root signature");
-
-    let our_pubkey = cfg.operator_table.pov_btc_key();
-    for (musig2_pubkey, nonce) in nonces.into_iter().filter(|(pk, _)| *pk != our_pubkey) {
-        info!(%musig2_pubkey, %deposit_request_txid, "loading nonce");
-        s2_client
-            .put_nonce(prevout, musig2_pubkey.to_x_only_pubkey(), nonce)
-            .await
-            .inspect_err(|e| {
-                error!(
-                    %deposit_request_txid,
-                    ?e,
-                    "failed to load nonce for root"
-                );
-            })?
-    }
-
-    info!("getting partial root sig");
-    let partial = s2_client
-        .get_partial(prevout, sighash)
-        .await
-        .inspect_err(|e| {
-            error!(
-                %deposit_request_txid,
-                ?e,
-                "failed to get partial root sig"
-            );
-        })?;
-
-    info!(%deposit_request_txid, "publishing root signature");
-    msg_handler
-        .send_musig2_signatures(
-            SessionId::from_bytes(prevout.txid.as_raw_hash().to_byte_array()),
-            vec![partial],
-        )
-        .await;
-
-    Ok(())
-}
-
-async fn handle_publish_deposit(
-    musig: &MusigSessionManager,
-    tx_driver: &TxDriver,
-    deposit_tx: DepositTx,
-    partials: BTreeMap<secp256k1::PublicKey, PartialSignature>,
-) -> Result<(), ContractManagerErr> {
-    info!(deposit_txid=%deposit_tx.compute_txid(), "executing duty to publish deposit");
-
-    let prevout = deposit_tx
-        .psbt()
-        .unsigned_tx
-        .input
-        .first()
-        .unwrap()
-        .previous_output;
-
-    for (pk, partial) in partials {
-        musig
-            .put_partial(prevout, pk.to_x_only_pubkey(), partial)
-            .await?;
-    }
-
-    let sig = musig.get_signature(prevout).await?;
-    let schnorr_sig = schnorr::Signature::from_slice(&sig.serialize())
-        .expect("must be a valid schnorr signature");
-    let taproot_sig = taproot::Signature {
-        signature: schnorr_sig,
-        sighash_type: TapSighashType::All,
-    };
-
-    let mut sighasher = SighashCache::new(deposit_tx.psbt().unsigned_tx.clone());
-
-    let deposit_tx_witness = sighasher.witness_mut(0).expect("must have first input");
-    deposit_tx_witness.push(taproot_sig.to_vec());
-
-    if let TaprootWitness::Script {
-        script_buf,
-        control_block,
-    } = &deposit_tx.witnesses()[0]
-    {
-        deposit_tx_witness.push(script_buf.to_bytes());
-        deposit_tx_witness.push(control_block.serialize());
-    }
-
-    let tx = sighasher.into_transaction();
-
-    info!(txid = %tx.compute_txid(), "broadcasting deposit tx");
-    tx_driver
-        .drive(tx, 0)
-        .await
-        .expect("deposit tx should get confirmed");
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn handle_publish_deposit_setup(
-    cfg: &ExecutionConfig,
-    s2_client: &SecretServiceClient,
-    db: &SqliteDb,
-    wallet: &RwLock<OperatorWallet>,
-    p2p_handle: &MessageHandler,
-    tx_driver: &TxDriver,
-    deposit_txid: Txid,
-    deposit_idx: u32,
-    stake_chain_inputs: StakeChainInputs,
-) -> Result<(), ContractManagerErr> {
-    info!(%deposit_txid, "executing duty to publish deposit setup");
-
-    let pov_idx = cfg.operator_table.pov_idx();
-    let scope = Scope::from_bytes(deposit_txid.as_raw_hash().to_byte_array());
-    let operator_pk = s2_client.general_wallet_signer().pubkey().await?;
-
-    let wots_client = s2_client.wots_signer();
-    /// VOUT is static because irrelevant so we're just gonna use 0
-    const VOUT: u32 = 0;
-    // withdrawal_fulfillment uses index 0
-    let withdrawal_fulfillment = Wots256PublicKey::from_flattened_bytes(
-        &wots_client
-            .get_256_public_key(deposit_txid, VOUT, 0)
-            .await?,
-    );
-    const NUM_FQS: usize = NUM_U256;
-    const NUM_PUB_INPUTS: usize = NUM_PUBS;
-    const NUM_HASHES: usize = NUM_HASH;
-    let public_inputs_ftrs: [_; NUM_PUB_INPUTS] =
-        std::array::from_fn(|i| wots_client.get_256_public_key(deposit_txid, VOUT, i as u32));
-    let fqs_ftrs: [_; NUM_FQS] = std::array::from_fn(|i| {
-        wots_client.get_256_public_key(deposit_txid, VOUT, (i + NUM_PUB_INPUTS) as u32)
-    });
-    let hashes_ftrs: [_; NUM_HASHES] =
-        std::array::from_fn(|i| wots_client.get_128_public_key(deposit_txid, VOUT, i as u32));
-
-    let (public_inputs, fqs, hashes) = join3(
-        join_all(public_inputs_ftrs),
-        join_all(fqs_ftrs),
-        join_all(hashes_ftrs),
-    )
-    .await;
-
-    info!(%deposit_txid, %deposit_idx, "constructing wots keys");
-    let public_inputs = public_inputs
-        .into_iter()
-        .map(|result| result.map(|bytes| Wots256PublicKey::from_flattened_bytes(&bytes)))
-        .collect::<Result<_, _>>()?;
-    let fqs = fqs
-        .into_iter()
-        .map(|result| result.map(|bytes| Wots256PublicKey::from_flattened_bytes(&bytes)))
-        .collect::<Result<_, _>>()?;
-    let hashes = hashes
-        .into_iter()
-        .map(|result| result.map(|bytes| Wots128PublicKey::from_flattened_bytes(&bytes)))
-        .collect::<Result<_, _>>()?;
-
-    let wots_pks = WotsPublicKeys::new(withdrawal_fulfillment, public_inputs, fqs, hashes);
-
-    // this duty is generated when this operator not only when a deposit request is observed
-    // but also when nagged by other operators.
-    // to avoid creating a new stake input, we first check the database.
-    info!(%deposit_txid, %deposit_idx, "checking if deposit data already exists");
-    if let Ok(Some(stake_data)) = db.get_stake_data(pov_idx, deposit_idx).await {
-        info!(%deposit_txid, %deposit_idx, "broadcasting deposit setup message from db");
-        let stakechain_preimg_hash = stake_data.hash;
-        let funding_outpoint = stake_data.operator_funds;
-
-        p2p_handle
-            .send_deposit_setup(
-                deposit_idx,
-                scope,
-                stakechain_preimg_hash,
-                funding_outpoint,
-                operator_pk,
-                wots_pks,
-            )
-            .await;
-
-        return Ok(());
-    }
-
-    info!(%deposit_txid, %deposit_idx, "constructing deposit setup message");
-    let StakeChainInputs {
-        stake_inputs,
-        pre_stake_outpoint,
-        ..
-    } = stake_chain_inputs;
-
-    info!(%deposit_txid, %deposit_idx, "querying for preimage");
-    let stakechain_preimg = s2_client
-        .stake_chain_preimages()
-        .get_preimg(
-            pre_stake_outpoint.txid,
-            pre_stake_outpoint.vout,
-            deposit_idx,
-        )
-        .await?;
-
-    let stakechain_preimg_hash = sha256::Hash::hash(&stakechain_preimg);
-
-    // check if there's a funding outpoint already for this stake index
-    // otherwise, find a new unspent one from operator wallet and filter out all the
-    // outpoints already in the db
-
-    info!(%deposit_txid, %deposit_idx, "fetching funding outpoint for the stake transaction");
-    let ignore = stake_inputs
-        .iter()
-        .map(|input| input.operator_funds.to_owned())
-        .collect::<HashSet<OutPoint>>();
-
-    let mut wallet = wallet.write().await;
-    info!("syncing wallet before fetching funding utxos for the stake");
-
-    match wallet.sync().await {
-        Ok(()) => info!("synced wallet successfully"),
-        Err(e) => error!(?e, "could not sync wallet but proceeding regardless"),
-    }
-
-    info!(?ignore, "claiming funding utxos");
-    let funding_op = wallet.claim_funding_utxo(|op| ignore.contains(&op));
-
-    let funding_utxo = match funding_op {
-        FundingUtxo::Available(outpoint) => outpoint,
-        FundingUtxo::ShouldRefill { op, left } => {
-            info!("refilling stakechain funding utxos, have {left} left");
-
-            let psbt = wallet.refill_claim_funding_utxos(FeeRate::BROADCAST_MIN)?;
-            finalize_claim_funding_tx(s2_client, tx_driver, wallet.general_wallet(), psbt).await?;
-
-            op
-        }
-        FundingUtxo::Empty => {
-            // The first time we run the node, it may be the case that the wallet starts off
-            // empty.
-            //
-            // For every case afterwards, we should receive a `ShouldRefill` message before
-            // the wallet is actually empty.
-            let psbt = wallet.refill_claim_funding_utxos(FeeRate::BROADCAST_MIN)?;
-            finalize_claim_funding_tx(s2_client, tx_driver, wallet.general_wallet(), psbt).await?;
-
-            let funding_utxo = wallet.claim_funding_utxo(|op| ignore.contains(&op));
-
-            match funding_utxo {
-                FundingUtxo::Available(outpoint) => outpoint,
-                _ => panic!("aaaaa no funding utxos available even after refill"),
-            }
-        }
-    };
-
-    info!(%deposit_txid, %deposit_idx, "constructing wots public keys for withdrawal fulfillment");
-    let withdrawal_fulfillment_pk = std::array::from_fn(|i| wots_pks.withdrawal_fulfillment[i]);
-
-    let stake_data = StakeTxData {
-        operator_funds: funding_utxo,
-        hash: stakechain_preimg_hash,
-        withdrawal_fulfillment_pk: strata_bridge_primitives::wots::Wots256PublicKey(
-            withdrawal_fulfillment_pk,
-        ),
-    };
-
-    info!(%deposit_txid, %deposit_idx, "adding stake data to the database");
-    debug!(%deposit_txid, %deposit_idx, ?stake_data, "adding stake data to the database");
-
-    db.add_stake_data(pov_idx, deposit_idx, stake_data).await?;
-
-    info!(%deposit_txid, %deposit_idx, "broadcasting deposit setup message");
-    p2p_handle
-        .send_deposit_setup(
-            deposit_idx,
-            scope,
-            stakechain_preimg_hash,
-            funding_utxo,
-            operator_pk,
-            wots_pks.clone(),
-        )
-        .await;
-
-    Ok(())
 }
