+++ crates/duty-tracker/src/executors/prelude.rs
@@ -1,0 +1,3 @@
+//! Re-exports specific executors for deposits and withdrawals for convenience.
+
+pub(crate) use super::{deposit::*, optimistic_withdrawal::*};
